[HOTSPOT-188] refactor: 폴더 이름 변경
